+++ modules/genealogy-dna-filament/src/Resources/DnaGroupResource/Pages/EditDnaGroup.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\Genealogy\Dna\Filament\Resources\DnaGroupResource\Pages;
+
+use Filament\Actions\DeleteAction;
+use Filament\Resources\Pages\EditRecord;
+use Liberu\Genealogy\Dna\Filament\Resources\DnaGroupResource;
+
+final class EditDnaGroup extends EditRecord
+{
+    protected static string $resource = DnaGroupResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [DeleteAction::make()];
+    }
+}